Moves configuration into separate file
